Laba 3.1.7

+++ app/Http/Controllers/ArticleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Article;
 use Illuminate\Http\Request;
 use App\Models\Comment;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
@@ -23,6 +24,7 @@ class ArticleController extends Controller
      */
 	public function create()
 	{
+        Gate::authorize('create', [self::class]);
 		return view('article.create_article');
 	}
 
@@ -31,6 +33,7 @@ class ArticleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', [self::class]);
         $request->validate([
             'name'=>'required',
             'short_desc'=>'required|min:6',
@@ -62,6 +65,7 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
+        Gate::authorize('update', [self::class, $article]);
         return view('article.edit_article', ['article' => $article]);
     }
 
@@ -70,6 +74,8 @@ class ArticleController extends Controller
      */
     public function update(Request $request, Article $article)
     {
+        Gate::authorize('update', [self::class, $article]);
+
         $request->validate([
             'name'=>'required',
             'short_desc'=>'required|min:6',
@@ -90,6 +96,7 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        Gate::authorize('delete', [self::class, $article]);
         $article->delete();
         return redirect('/article');
     }
