+++ resources/views/article/one_article.blade.php
@@ -14,14 +14,16 @@
         <p>{{ $article->desc }}</p>
         <p>{{ $article->date }}</p>
 
-        <div class="button-box">
-            <a href="/article/{{$article->id}}/edit" class="button button_blue">Редактировать</a>
-            <form class="form-for-button" action="/article/{{ $article->id }}" method="POST"> 
-                @csrf
-                @method('DELETE')
-                <button class="button button_red" type="submit">Удалить</button>
-            </form>
-        </div>
+        @can('update')
+            <div class="button-box">
+                <a href="/article/{{$article->id}}/edit" class="button button_blue">Редактировать</a>
+                <form class="form-for-button" action="/article/{{ $article->id }}" method="POST"> 
+                    @csrf
+                    @method('DELETE')
+                    <button class="button button_red" type="submit">Удалить</button>
+                </form>
+            </div>
+        @endcan
     </div>
 </section>
 <section class="comment-section">
@@ -44,13 +46,16 @@
     <div class="comments-container">
         @foreach ($comments as $comment)
             <div class="comment">
+                <p class="comment__author">{{ $comment->getAuthorName() }}</p>
                 <p class="comment__title">{{ $comment->title }}</p>
                 <p class="comment__text">{{ $comment->text }}</p>
                 <p class="comment__date">{{ $comment->created_at }}</p>
-                <div class="button-box">
-                    <a href="/comment/edit/{{ $comment->id }}" class="button button_blue">Редактировать</a>
-                    <a href="/comment/delete/{{ $comment->id }}" class="button button_red">Удалить</a>
-                </div>
+                @can('comment', $comment)
+                    <div class="button-box">
+                        <a href="/comment/edit/{{ $comment->id }}" class="button button_blue">Редактировать</a>
+                        <a href="/comment/delete/{{ $comment->id }}" class="button button_red">Удалить</a>
+                    </div>
+                @endcan
             </div>
         @endforeach
     </div>
